+++ crates/amaru-ledger/src/rules/transaction/phase_one/fixture.rs
@@ -197,7 +197,9 @@ pub(super) enum Predicate {
     OutsideValidityIntervalUTxO,
     ScriptsNotPaidUTxO,
     TooManyCollateralInputs,
+    ProposalReturnAccountDoesNotExist,
     TreasuryWithdrawalReturnAccountsDoNotExist,
+    TreasuryWithdrawalsAllZeros,
     DelegateeDRepNotRegistered,
     DelegateeStakePoolNotRegistered,
     DRepAlreadyRegistered,
@@ -264,6 +266,12 @@ impl From<PhaseOneError> for Predicate {
             },
             PhaseOneError::Proposals(InvalidProposals::TreasuryWithdrawalReturnAccountsDoNotExist(_)) => {
                 Predicate::TreasuryWithdrawalReturnAccountsDoNotExist
+            }
+            PhaseOneError::Proposals(InvalidProposals::TreasuryWithdrawalsAllZeros) => {
+                Predicate::TreasuryWithdrawalsAllZeros
+            }
+            PhaseOneError::Proposals(InvalidProposals::ProposalReturnAccountDoesNotExist(_)) => {
+                Predicate::ProposalReturnAccountDoesNotExist
             }
             PhaseOneError::ValueNotPreserved(_) => Predicate::ValueNotConservedUTxO,
             PhaseOneError::Certificates(InvalidCertificates::StakeCredentialInvalidPoolDelegation(ref e)) => match e {